requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
datasets
peft
pyarrow>=15.0.0
pybind11
pylatexenc
pylint==3.3.6
ray[default]
torchdata
# vllm==0.6.3.post1
tensordict<0.6
transformers
vllm
wandb
